+++ CHANGELOG.rst
@@ -1,9 +1,15 @@
-===========
+***********
 Change Log
-===========
+***********
+
+0.2 (2014-07-28)
+=================
+
+   * Added support for optional fields
+   * Added support for default arguments for fields
+   * Option to retrieve segments by name
 
 0.1 (2014-06-16)
 =================
 
    * Babel supports requirements of hypothetical v2 API.
-